Got Lighting sort-of-but-not-really working
Made a sample light and surface for the BSP map, and produced the
normals of the map's surfaces. There is one problem, however, and that
is with OpenGL's default vertex shader, which ignores the positions of
the map's vertices when doing lighting calculations. The solution to
this will be implementing our own Shader object and writing a per-vertex
or per-fragment shader.